Process: simplified 2-to-4 implementation, and dropped intermediate layer of central phase space generator
